Push branch.

+++ script/src/main/kotlin/script/packet/disassembler/MovementDisassembler.kts
@@ -1,8 +1,8 @@
 package script.packet.disassembler
 
-import io.ktor.utils.io.core.readUByte
 import xlitekt.game.packet.MovementPacket
 import xlitekt.game.packet.disassembler.onPacketDisassembler
+import xlitekt.shared.buffer.readUByte
 import xlitekt.shared.buffer.readUShortAdd
 import xlitekt.shared.buffer.readUShortLittleEndianAdd
 
@@ -12,7 +12,7 @@ import xlitekt.shared.buffer.readUShortLittleEndianAdd
  */
 onPacketDisassembler(opcodes = intArrayOf(57, 58), size = -1) {
     MovementPacket(
-        movementType = readUByte().toInt(),
+        movementType = readUByte(),
         destinationZ = readUShortLittleEndianAdd(),
         destinationX = readUShortAdd()
     )
